requirements.txt
djangorestframework
python-decouple
psycopg2-binary
Pillow
